project_test_1/orderbook-api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
pydantic>=2.11.0
pydantic-settings==2.6.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
